Code fix

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned int nodes = 0;
+	size_t nodes = 0;
 
 	while (h != 0)
 	{
